backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
Pillow>=9.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
aiofiles>=23.0.0
httpx>=0.24.0
